chore: updated file name on useCases and created index import
