+++ src/useCases/index.ts
@@ -1,0 +1,8 @@
+export * from "./cliente/clienteUseCase";
+export * from "./cliente/dto";
+
+export * from "./produto/produtoUseCase";
+export * from "./produto/dto";
+
+export * from "./pedido/pedidoUseCase";
+export * from "./pedido/dto";
