+++ tests/useCases/pedido/index.test.ts
@@ -3,7 +3,7 @@ import { StatusPedidoEnum, Pedido } from "entities/pedido";
 import { Produto, CategoriaEnum } from "entities/produto";
 import { PedidoGateway } from "interfaces/gateways/pedidoGateway.interface";
 import { ProdutoGateway } from "interfaces/gateways/produtoGateway.interface";
-import { PedidoUseCase } from "useCases/pedido";
+import { PedidoUseCase } from "useCases";
 import { Email, Cpf } from "valueObjects";
 
 const mockClienteDTO = {
